Move all DOM related function to domUpdates.js.
