+++ src/index.html
@@ -8,14 +8,14 @@
       <img src="./images/ExcursiOnward-logo.png" alt="ExcursiOnward: where adventure starts" height="50px" width="50px">
       <h1>ExcursiOnward</h1>
     </header>
-    <section class="login-dashboard">
+    <section class="login-dashboard hidden">
       <label for="login-user-name">Username</label>
       <input type="text" name="login-user-name" id="login-user-name">
       <label for="login-password">Password</label>
       <input type="text" name="login-password" id="login-password">
       <button class="login-button">Log In!</button>
     </section>
-    <main class="dashboard hidden">
+    <main class="dashboard">
       <section class="dashboard-trips">
         <h1 class="dashboard-greeting"></h1>
         <article class="dashboard-trips-past">
