+++ src/index.js
@@ -4,7 +4,6 @@ import './images/ExcursiOnward-logo.png'
 import './images/map-background.jpg'
 
 import Trip from './Trips';
-import Destination from './Destination'
 import Traveler from './Traveler'
 import Agency from './Agency'
 
@@ -16,8 +15,12 @@ import {
   updateTrip,
   deleteTrip
 } from './utility.js';
+
+import {
+  domUpdates
+} from './domUpdates';
+
 let today = new Date().toISOString().slice(0,10).replaceAll("-", "/")
-let lastYear = new Date(new Date().setFullYear(new Date().getFullYear() - 1)).toISOString().slice(0,10).replaceAll("-", "/");
 
 // login
 const loginScreen = document.querySelector('.login-dashboard');
@@ -27,19 +30,6 @@ const loginButton = document.querySelector('.login-button');
 
 // dashboard
 const dashboard = document.querySelector('.dashboard')
-const dashboardGreeting = document.querySelector('.dashboard-greeting')
-const destinationList = document.querySelector('.destination-list')
-const pastTripList = document.querySelector('.past-trips')
-const upcomingTripList = document.querySelector('.future-trips')
-const currentTripList = document.querySelector('.current-trips')
-
-// highlight
-const highlight = document.querySelector('.destination-preview')
-const highlightTitle = document.querySelector('.destination-preview-title')
-const highlightFlight = document.querySelector('.flight-cost')
-const hightlightDaily = document.querySelector('.expense-cost')
-const highlightImage = document.querySelector('.destination-preview-image')
-
 
 // trip request selectors
 const tripStartDate = document.querySelector('.start-date')
@@ -51,14 +41,14 @@ const estimateButton = document.querySelector('.get-estimate')
 const tripPreviewTitle = document.querySelector('.trip-preview-title');
 const tripPreviewData = document.querySelector('.trip-preview-data');
 const tripPreviewImage = document.querySelector('.trip-preview-image');
-const yearCost = document.querySelector('.dashboard-year-cost');
+
 
 // let travelerID = (Math.ceil(Math.random() * 49))
 let travelerID
 
 const pageLoad = () => {
 
-  const travelerResults = getTraveler(travelerID)
+  const travelerResults = getTraveler(44)
   const tripsResults = getTrips()
   const placeResults = getDestinations()
     
@@ -76,74 +66,19 @@ const pageLoad = () => {
         traveler.isThisMyTrip(trip)
       })
       values[2].forEach(destination => {
-        addToDestinationList(destination)
+        domUpdates.addToDestinationList(destination)
       })
-      displayHighlight(values[2])
-      displayTravelerName(traveler);
-      displayUserTrips(traveler);
-      displayYearlyCost(traveler);
+      domUpdates.displayHighlight()
+      domUpdates.displayTravelerName(traveler);
+      domUpdates.displayUserTrips(traveler);
+      domUpdates.displayYearlyCost(traveler);
     })
-}
-
-const displayTravelerName = (traveler) => {
-  dashboardGreeting.innerText = "Hello, " + (traveler.returnFirstNameLastInitial())
-}
-
-const addToDestinationList = (destination) => {
-  let option = document.createElement('option');
-  destinationList.appendChild(option);
-  option.innerText = `${destination.name}`
-  option.setAttribute('value', `${destination.id}`)
-}
-
-const displayHighlight = () => {
-  getDestinations()
-  .then(places => {
-    let highlightPlace = places.find(place => {
-      return place.id === (Math.floor(Math.random() * 50))
-    })
-    highlightTitle.innerText = `${highlightPlace.name}`
-    highlightFlight.innerText = `Flights start at $${highlightPlace.costPerPerson}`
-    hightlightDaily.innerText = `Stay for as little as $${highlightPlace.costPerDay} a day!`
-    highlightImage.setAttribute('src', highlightPlace.image)
-    highlightImage.setAttribute('alt', highlightPlace.altText)
-  })
-}
-
-const showTrip = (parent, trip) => {
-    let li = document.createElement('li');
-    parent.appendChild(li);
-    li.innerText = `${trip.destinationData.name}, Date: ${trip.date}, ${trip.duration} days, Cost: $${trip.calculateTotalCost().toFixed(2)}, Status: ${trip.status}, TripID: ${trip.id}`
-};
-
-const displayUserTrips = (traveler) => {
-  traveler.trips.forEach(trip => {
-    if (trip.happeningData === 'past') {
-      showTrip(pastTripList, trip);
-    } else if (trip.happeningData === 'upcoming') {
-      showTrip(upcomingTripList, trip);
-    } else if (trip.happeningData === 'current') {
-      showTrip(currentTripList, trip);
-    }
-  })
 }
 
 const findDuration = (start, end) => {
   let startDate = new Date(start);
   let endDate = new Date(end);
   return (endDate.getDate()) - (startDate.getDate())
-}
-
-const displayEstimate = (newTrip, destinationData) => {
-  tripPreviewTitle.innerText = `Your trip to ${destinationData.name}`
-  tripPreviewData.innerHTML = `
-  <p class="destination">Destination ${destinationData.name}</p>
-  <p class="leaving">Departing on: ${newTrip.date}</p>
-  <p class="duration">Duration ${newTrip.duration}</p>
-  <p class="total-cost">Cost: $${newTrip.calculateTotalCost()}</p>
-  <button class="book-trip">Book It!</button>`
-  tripPreviewImage.setAttribute('src', destinationData.image)
-  tripPreviewImage.setAttribute('alt', destinationData.altText)
 }
 
 const returnTripEstimate = (event) => {
@@ -166,18 +101,15 @@ const returnTripEstimate = (event) => {
       "suggestedActivities": ['do stuff'],
     }, destinationData)
     if (event.target.classList.contains('get-estimate')) {
-      displayEstimate(newTrip, destinationData)
+      domUpdates.displayEstimate(newTrip, destinationData)
     } else if (event.target.classList.contains('book-trip')) {
     addTrip(newTrip)
-    showTrip(upcomingTripList, newTrip)
+    domUpdates.showTrip(upcomingTripList, newTrip)
     }
   })
 }
 
-const displayYearlyCost = (traveler) => {
-  console.log(traveler.findYearlyTravelCost(lastYear))
-  yearCost.innerText = `You have spent $${traveler.findYearlyTravelCost(lastYear).toFixed(2)} in the last year.`
-}
+
 
 const login = (e) => {
   e.preventDefault();
